+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.8.1] - 2023-09-29
+
 * Fix: support `token_data.to_s` when using mocked JWTs
 
 ## [2.8.0] - 2023-09-29
@@ -112,7 +114,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add a changelog
 * Setup automated gem publishing
 
-[Unreleased]: https://github.com/zaikio/zaikio-directory-models/compare/v2.8.0...HEAD
+[Unreleased]: https://github.com/zaikio/zaikio-directory-models/compare/v2.8.1...HEAD
+[2.8.1]: https://github.com/zaikio/zaikio-directory-models/compare/v2.8.0...v2.8.1
 [2.8.0]: https://github.com/zaikio/zaikio-directory-models/compare/v2.7.0...v2.8.0
 [2.7.0]: https://github.com/zaikio/zaikio-directory-models/compare/v2.7.0...v2.7.0
 [2.6.0]: https://github.com/zaikio/zaikio-directory-models/compare/v2.5.0...v2.6.0